requirements.txt
# Interface
PyGObject>=3.48.0
# Scraping
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
# Visão computacional
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
# IA local
httpx>=0.25.0
# Utilitários
python-dotenv>=1.0.0
# Dev
pytest>=8.0.0
pytest-cov>=4.1.0
ruff>=0.3.0
pre-commit>=3.5.0
